+++ branches/dev_curriculum/JavaSource/org/unitime/timetable/solver/studentsct/StudentSectioningDatabaseLoader.java
@@ -259,8 +259,8 @@ public class StudentSectioningDatabaseLoader extends StudentSectioningLoader {
                     (room instanceof Room? ((Room)room).getBuilding().getUniqueId() : null),
                     0,
                     room.getCapacity().intValue(),
-                    room.getCoordinateX().intValue(),
-                    room.getCoordinateY().intValue(),
+                    room.getCoordinateX(),
+                    room.getCoordinateY(),
                     room.isIgnoreTooFar().booleanValue(),
                     null);
             rooms.addElement(roomLocation);
